backend/requirements.txt
# requirements.txt
Django==4.2.5
djangorestframework==3.14.0
django-cors-headers==4.2.0
django-filter==23.2
drf-spectacular==0.26.5
psycopg2-binary==2.9.7
celery==5.3.1
redis==4.6.0
django-redis==5.4.0
python-dotenv==1.0.0
Pillow==10.0.0
jsonschema==4.23.0
python-json-logger==2.0.7
requests==2.32.3